+++ node.gyp
@@ -280,8 +280,8 @@
         'src/tty_wrap.h',
         'src/tcp_wrap.h',
         'src/udp_wrap.h',
-        'src/req-wrap.h',
-        'src/req-wrap-inl.h',
+        'src/req_wrap.h',
+        'src/req_wrap-inl.h',
         'src/string_bytes.h',
         'src/stream_base.h',
         'src/stream_base-inl.h',
