+++ src/cares_wrap.cc
@@ -24,7 +24,7 @@
 #include "async_wrap-inl.h"
 #include "env-inl.h"
 #include "node.h"
-#include "req-wrap-inl.h"
+#include "req_wrap-inl.h"
 #include "util-inl.h"
 #include "uv.h"
