+++ src/connect_wrap.h
@@ -4,7 +4,7 @@
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #include "env.h"
-#include "req-wrap.h"
+#include "req_wrap.h"
 #include "async_wrap.h"
 #include "v8.h"
