+++ src/req_wrap-inl.h
@@ -3,7 +3,7 @@
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#include "req-wrap.h"
+#include "req_wrap.h"
 #include "async_wrap-inl.h"
 #include "env-inl.h"
 #include "util-inl.h"
