+++ src/stream_base.h
@@ -5,7 +5,7 @@
 
 #include "env.h"
 #include "async_wrap.h"
-#include "req-wrap-inl.h"
+#include "req_wrap-inl.h"
 #include "node.h"
 #include "util.h"
