+++ src/tty_wrap.cc
@@ -25,7 +25,7 @@
 #include "handle_wrap.h"
 #include "node_buffer.h"
 #include "node_wrap.h"
-#include "req-wrap-inl.h"
+#include "req_wrap-inl.h"
 #include "stream_wrap.h"
 #include "util-inl.h"
