+++ src/udp_wrap.h
@@ -27,7 +27,7 @@
 #include "async_wrap.h"
 #include "env.h"
 #include "handle_wrap.h"
-#include "req-wrap-inl.h"
+#include "req_wrap-inl.h"
 #include "uv.h"
 #include "v8.h"
